Implement TasksDragAndDropContext
- Create TasksDragAndDropContext for Drag and Drop functionality
- Enhance TasksWeekItemGroup by replacing the existing DnD context with TasksDragAndDropContext
